Update KSM version to match what we deploy
The versions check was broken in the CI until recently and operand
versions in jsonnet/versions.yaml were not validated properly. Now that
the CI job is fixed, all version checks are starting to fail for newly
created PRs.

This commit regenerates the versions to make sure they are correct.