Set arguments in a deterministic order
When setting arguments for the prometheus-operator deployment, we
flatten them out into a map for easier processing before merging them
into a list again. This causes the arguments to always end up in a
different order and re-triggers a new rollout of the prometheus-operator
deployment.

This PR changes the implementation to preserve the order of the
prometheus-operator arguments so that new pods are rolled out only when
the arguments change.

Fixes commit 857f03cd03af807912d837c0cde3ff18213efafe