Revert of revert "Configure prometheus operator TLS ..." with argument order fix (#1409)
* Revert "Revert "Configure prometheus operator TLS based on the cluster APIServer config (#1380)""

This reverts commit 0b31a1eca555210a5d6d48e73c4bf5096c282ea8.

* Set arguments in a deterministic order

When setting arguments for the prometheus-operator deployment, we
flatten them out into a map for easier processing before merging them
into a list again. This causes the arguments to always end up in a
different order and re-triggers a new rollout of the prometheus-operator
deployment.

This PR changes the implementation to preserve the order of the
prometheus-operator arguments so that new pods are rolled out only when
the arguments change.

Fixes commit 857f03cd03af807912d837c0cde3ff18213efafe