prometheus-operator: explicitly depend on the ca key of the configmap
This prevents the pod/container to start before the service-ca-operator
has populated the respective configmap. Otherwise it can happen that a
container starts (since the configmap is present) but crashes
immediately due to errors in parsing the pem content (since the key in
the configmap is still empty). See also https://kubernetes.io/docs/tasks/configure-pod-container/configure-pod-configmap/#restrictions

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>