jsonnet: fix usage of 'config.openshift.io/inject-trusted-cabundle': 'true'
This label causes the cluster-network-operator to inject a trusted CA
bundle into the respective configmap. This commit refactors the
configmap create into a function and removes the default empty
ca-bundle.crt key from data. The cluster-network-operator will create
this key.
Pods adding this configmap as a volume should explicitly list this key
in `items` so that they won't get started before the CA data is
available. An empty key allows the pod to start and containers to try
and parse CA data that is empty.
This overlaps with cdb0980d713adcaa1a63d26bddb8b665ae7b6129 and might
not change anything in the deployment order, but its nontheless the
correct way and should yield less surprises on later code changes.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>