Bug 1995924: Set Upgradeable: false when HA workloads are incorrectly spread (#1431)
* Merge pull request #1330 from dgrisonnet/ha-upgradeable

Bug 1995924: Set `Upgradeable: false` when HA workloads are incorrectly spread

* pkg/rebalancer: sort resources for deletion

The resources that are marked for deletion are sorted by their PVC
creation timestamp, from the newest to the oldest to make the deletion
consistent.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>

* pkg: improve rebalancer logging

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>

* pkg/rebalancer: sort based on pod names

Sort PVCs by their creation timestamps, from the newest to the oldest to
make sure that the oldest PVC is retained in case all of them are
annotated. If some PVCs have the same creation timestamp, they will be
sorted based on their pod name.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>

* pkg/rebalancer: split annotation removal

Add EnsurePVCsAreNotAnnoted function that makes sure that none of the
PVCs of the given workload have the
openshift.io/cluster-monitoring-drop-pvc annotation after the
rebalancing is done. In case one of the PVC has the annotation, it will
be removed to prevent deleting the PVC in a future cycle.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>

* pkg/rebalancer: sort resources to delete by age

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>

* test/e2e: fix framework

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>

Co-authored-by: OpenShift Merge Robot <openshift-merge-robot@users.noreply.github.com>