*: enable hard-anti affinity and PDB for Alertmanager
This change introduces hard pod anti-affinity rules and pod disruption
budgets for Alertmanager to ensure the maximum availability for the
Alertmanager cluster in the event of nodes going down (either due to
upgrades or unexpected outages). The cluster monitoring operator updates
the `Upgradeable` condition to false when it detects that the pods
aren't correctly spread to ensure that an upgrade only happens in safe
configurations.

The change also decreases the number of Alertmanager replicas from 3 to
2 to be consistent with the other monitoring components as well as the
  HA conventions stating that in general OpenShift component should run
with a replica count of 2 [1]. In addition with 3 replicas, it is
impossible to enable hard anti-affinity on nodes since 2 worker nodes is
a supported deployment for OCP.

The initial idea of running 3 replicas was to guarantee the replication
of data (silences + notifications) during pod roll-outs even if the user
didn't configure persistent storage. However given that no pod
disruption budget were defined, there was no guarantee that Kubernetes
would always keep one Alertmanager pod running.  With hard anti-affinity
and PDB, we are now sure that at least one Alertmanager pod is kept
running. And we are also setting up a startup probe that waits for at
least 20 seconds meaning that Kubernetes should wait for about 20
seconds after a new Alertmanager pod is running before considering
rolling out the next one. This interval of time should be more than
enough for the new Alertmanager to synchronize its data from the older
peer.

[1]
https://github.com/openshift/enhancements/blob/master/CONVENTIONS.md#high-availability

Signed-off-by: Simon Pasquier <spasquie@redhat.com>