Enable PDB labels metrics
Though PDB labels are enabled already it is not working yet due to a
typo bug in upstream KSM. Usually allowlist resource name should be in
plural form, but incase of PDB KSM is using singular form of resource name
which is incorrect.

This commit is a workaround to enable PDB by passing singular form of
PDB resource name. This commit should be reverted once the upstream KSM
fixes the typo.

Signed-off-by: Arunprasad Rajkumar <arajkuma@redhat.com>