pkg/manifests: increase retries for  Alertmanager startup probe
On some platforms, it may take more than 20s for Alertmanager to resolve
the peer names in which case, the startup probe fails because the web
server isn't ready and kubelet will continuously restart the container.

This change increases the number of retries to 4 with a period of 10
seconds meaning that it can take up to 1 minute (20s + 4 x 10s) for the
Alertmanager's HTTP server to be running.

See https://bugzilla.redhat.com/show_bug.cgi?id=2037073 for details.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>