prometheus: remove ui access
Type: feature removal

Problem: We expose the full prometheus http api to authorized users
through a route, i.e. its externally available. We want to remove access
to the UI for a reduced support footprint, while maintaining access to
`/api` and `/federate`.

Solution: Alter the existing Route object to be limited to `/api` and
create a new Route object to expose `/federate`.
For the externalUrl configuration (used in backlinks in alerts for
example) the console route is now used.
Another solution considered was to alter the oauth-proxy config. This is
discarded due to oauth-proxy feature not alligning as needed and any
change there also impacting the Service object that exposes the
deployment cluster-internal.

Issue: https://issues.redhat.com/browse/MON-1631

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>