pkg/tasks: Don't manage alert relabel configs secret
This updates the internal naming for the asset containing alert
relabel configs to be more consistent, and removes the bit from the
Prometheus task that manages the secret.  This is in preperation for
having the relabel configs managed by a new controller, which allows
for them to be changed dynamically.