test: account for pdb and Prometheus' staleness period
Problem: We recently introduced pod disruption budgets and improved
availability on Prometheus. This is reflected in the up metric for
Prometheus. By virtue of Prometheus staleness mechanism (it will return
samples up to 5m old for a series) the old query can appear to see 4
Prometheus instances.

Solution: Use the last_over_time aggregation with a 2m interval to only
look at the last sample in that interval. This effectively adjusts
Prometheus staleness period to 2m.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>