Properly deal with an empty console URL
Currently, when the console object exist but has an empty url (happens
e.G. during cluster creation), an invalid alertmanager statefulset is
produced that has an argument of `--web.external-url=monitoring`.

This is made worse by the fact that we wait up to five minutes for the
alertmanager pods to become ready, which will never happen with this config,
resulting in a lot of startup delay if the CMO happens to be quicker
than the console operator.

This also fixes this for Prometheus, where it isn't strictly required as
prometheus seems to accept `monitoring` as a valid `external-url`. Odds
are this validation will be added there too, so fix that while we are at
it.