Bug 2043518: set degraded and available status based on Prometheus pod status
NOTE: This patch addresses 2 related Bugs:
1. Bug 2043518: Better message in the CMO degraded/unavailable conditions when pods can't be scheduled
1. Bug 2039411: Monitoring operator reports unavailable=true while one Prometheus pod is ready

The patch sets the Degraded and Availability status of ClusterOperator
by inspecting Prometheus Pod status. To achieve this all tasks now
return a StateError that captures if the state of the system is
Degraded or Unavailable and its  associated reasons for the change.

Thus a task's Run method is now structured as follows

```go
func (t *Task) Run() *StateError { }
```
NOTE: This patch only modifies the Prometheus Task and separates our the
validation part into a different task which can later be reused to validate
minimal availability of Prometheus.

Thus, given the following monitoring config which uses a non-existent PVC

```yaml
  prometheusK8s:
    volumeClaimTemplate:
      metadata:
        name: prometheus-db
      spec:
        storageClassName: foo
        resources: { requests: {storage: 100Mi} }
```

It now correctly marks the ClusterOperator status as follows (formatted
for readability)

```
  Monitoring Stack is Unavailable.
  Error: updating prometheus-k8s
  status: unavailable, degraded;
  reasons:
    prometheus pod openshift-monitoring/prometheus-k8s-0 is pending due to Unschedulable
  : 0/6 nodes are available
  : 6 pod has unbound immediate PersistentVolumeClaims., prometheus pod openshift-monitoring/prometheus-k8s-1 is pending due to Unschedulable
  : 0/6 nodes are available
  : 6 pod has unbound immediate PersistentVolumeClaims
```

Signed-off-by: Sunil Thaha <sthaha@redhat.com>