Documentation/data-collection: Assign cluster-version metrics to Cincinnati team
The installer assignments are from a7146706b4 (manifests: update
owners for telemetry metrics, 2021-07-07, #1273).  Not clear to me why
that assignment was chosen then, but the Cincinnati team [1] (the
GitHub name of the OpenShift-core updates team) is in charge of the
cluster-version operator, and these are all metrics served by the CVO
[2].

cluster_installer leans heavily on installer-supplied data, so I'm
leaving that one co-owned by the installer team.

Also from a7146706b4 was the two 'owners' lines for cluster_installer.
I don't think OLM was intended to be an owner for this metric (and
they're separately listed as a consumer), so I've dropped their owner
line.

[1]: https://github.com/orgs/openshift/teams/openshift-team-cincinnati/members
[2]: https://github.com/openshift/cluster-version-operator/blob/fc7d096f5fcde2758dafd1ce968601cd774479f4/pkg/cvo/metrics.go#L65-L105