*: add standlone admission webhook
Before this commit, the validating webhook configurations for
PrometheusRule and AlertmanagerConfig resources leveraged the webhook
service of Prometheus operator. Because the operator doesn't implement
any leader election, it has to run as a single replica meaning that the
webhook service couldn't be highly-available (especially during
upgrades).

This change deploys the admission webhook as a standalone service with 2
replicas + hard anti-affinity + pod disruption budget. A follow-up PR
will switch the failure policy of the validating webhook configurations
for PrometheusRule and AlertmanagerConfig resources from 'Ignore' to
'Fail'.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>