Bug 2033575: use bearer token as fall-back authn method
When the TLS certificate used for authentication gets rotated, Prometheus
doesn't pick up the new certificate until the connection to the target is
re-established. Because Prometheus uses keep-alive HTTP connections, the
consequence is that the scrapes start failing after about 1 day and the
TargetDown alert fires. To resolve the alert, the cluster admin has to
restart the pods being reported as down.

To workaround the issue (and until Prometheus properly handles certificate
rotation), patch the service monitors in the openshift-monitoring and
openshift-user-workload-monitoring namespaces with fall-back authentication
method using the service account bearer token.

Details in:
https://github.com/prometheus/prometheus/issues/9512 (upstream)
https://bugzilla.redhat.com/show_bug.cgi?id=2033575

Signed-off-by: Simon Pasquier <spasquie@redhat.com>