Updates CreateOrUpdateDeployment Issue https://bugzilla.redhat.com/show_bug.cgi?id=2069068
problem: when reconciling the prometheus operator(PO) deployment sometimes
CMO would get an API error and it would delete and create the PO deployment
in a short span of time without giving time to the API server to reconcile.

solution: instead we should validate first that the deployment is
actually deleted as we do for CreateDeployment.