Refactors CreateRouteIfNotExists to CreateOrUpdateRoute Issue: https://bugzilla.redhat.com/show_bug.cgi?id=2084079
Problem: on OpenShift 4.11 the Prometheus route was updated with
a subpath /api, CMO was only able to create routes and not update them.
When users would update from a version before 4.11 to 4.11 the route
would not be updated.

Solution: allow CMO to Create or Update Routes similar to how it
interects with other k8s resources.