Removes etcd related dashboards from CMO Issue: https://bugzilla.redhat.com/show_bug.cgi?id=2089199
Problem: Currently CMO deploys a set of dashboards that report
information about etcd, these however do not make sense if etcd is not
deployed in the cluster (hypershift).

Solution: Move these dashboard to the etcd operator repo so that these
dashboards are only deployed to the cluster when etcd itself is running
there.