Bug 2100472: fix alert controllers when not in techpreview
When the TechPreview feature gate isn't enabled, the AlertingRule
and AlertRelabelConfig CRDs don't exist which triggers warning logs when
the operator wants to start the respective informers.

When not in techpreview mode, the controllers initialize the informers
with fake lister+watcher that return no items.

This commit also re-enables the CMO e2e tests which were accidentally
disabled.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>