Adds telemeter token hash to Deployment Issue: https://bugzilla.redhat.com/show_bug.cgi?id=2114721
Problem: Currently after a pull secret is changed/updated the telemeter
client starts failing to authenticate with telemeter server. This
happens because the client does not load the new secret unless it is
restarted. The salt in the secret is being continuosly updated which is
not correct.

Solution: Add a hash of the secret to the deployment so that when the
secret is rottated the deployment is restarted. Only update the salt
in the secret when the token is updated