pkg/config: add new config flag K8sPrometheusAdapter.DedicatedServiceMonitors.Enabled
When this config flag is set to true, CMO will load an additional
kubelet ServiceMonitor, that exposes metrics that prometheus-adapter
queries for pod CPU and RAM usage. Additionally the prometheus-adapter
config ConfigMap, which query these metrics, is loaded.
In order for the prometheus-adapter pods to load the new config, the
ConfigMap volume name needs to change. This required changing the
deployment to inject the ConfigMap name into the spec.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>