control-plane: use `/metrics/resource` endpoint for PA dedicated service monitor
In order to minimize resource overhead we want to use the
`/metrics/resource` endpoint of kubelet. This exposes fewer metrics, but
includes the ones we are interested in. Since kube-prometheus does not
provide jsonnet code for this endpoint we just re-use the
`/metrics/cadvisor` jsonnet code and change it as needed.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>