Bug OCPBUGS-47741: increase alertmanager's startupProbe failure threshold
On some clusters, during the initial boot, DNS resolution of services
may fail and this can cause Alertmanager pods to fail, which in turn
causes CMO to be in degraded state.

This patch fixes the issue by increasing the startupProbe's tolerance.

NOTE: The actual issue in the Bug[1] seems to be DNS related where
alertmanager fails to establish a quorum since its peers can't be
reached.

[1]: https://issues.redhat.com/browse/OCPBUGSM-47741
[1]: https://bugzilla.redhat.com/show_bug.cgi?id=2083226

Signed-off-by: Sunil Thaha <sthaha@redhat.com>