container_fs_.* metrics doesn't have container label after openshift/pull/1402 which caused regression for dashboards using container label since it gives empty datapoints for all queries using container label in queries containing container_fs_.* metrices.
Inorder to fix the dashboards added containerfsSelector in
kubernetes-monitoring/kubernetes-mixin/pull/737 and overrided the containerfsSelector
default value from  container!= to harmless id!= in this commit.
We really don't need that selector since we dropped the container label
in OCP for container_fs_.* metrices but we need to override the
selector for the queries involving container_fs_.* to work.

Signed-off-by: Jayapriya Pai <janantha@redhat.com>