Fixes externalURL field for Prometheus and Alertmanager
Issue: https://issues.redhat.com/browse/OCPBUGS-4181

Problem: currently when setting externalURL field CMO uses a package
which removes one of the two "/" after "https" this causes that when
user get alerts in their receivers that alerts have in the console URL
a total of 3 "/"

Solution: instead of using the "path" package to join the "monitoring"
path with the console URL use instead the net/url package. Add tests to
manifest_test.go to test this behaviour.

Signed-off-by: JoaoBraveCoding <jmarcal@redhat.com>