OCPBUGS-2141: compute doc link in PVC not configured message
Previously, when PVC is not configured for Prometheus, the CO monitoring
status message for Degraded condition included a link to the
documentation on configuring monitoring stack for version 4.8.

This commit fixes the link to point to doc for the current cluster
version and in case of failure to compute, the cluster-version, points
to the latest documentation.

JIRA: https://issues.redhat.com/browse/OCPBUGS-2141

Signed-off-by: Sunil Thaha <sthaha@redhat.com>