OCPBUGS-7391: wait for service CA secrets
The Prometheus adapter and control plane tasks would exit immediately if
they couldn't find the TLS secrets generated by the service CA operator.

With this change, the tasks are waiting up to 5 minutes for the secrets
to appear which should speed up a bit the reconciliation (no need to run
another attempt).

Signed-off-by: Simon Pasquier <spasquie@redhat.com>