jsonnet: Add prometheus container in UWM prometheus asset file
openshift#1824 increased the
startupProbe timeout from 15m to 1h for both platform and uwm
proemetheus. But it didn't take effect on UWM prometheus
because in assets file it didn't have a container named `prometheus`
on UWM prometheus pod so CMO never entered the code path
that was added in openshift#1824
since it uses strategic merge patch.

Hence this commit adds a `prometheus` container in
uwm prometheus jsonnet file so it gets added
in assets file.

Fixes #OCPBUGS-10690

Signed-off-by: Jayapriya Pai <janantha@redhat.com>