MON-3173: Remove INTERACTIVE mode from hack/generate-versions.sh to simplify.
Make hack/generate-versions.sh do nothing on PR that are not against the main branch as the components versions are only updated on the main branch (on release-xx branches, components kepp their versions when fixed)

Add a 'make check-versions' target to run the versions check. CI jobs will be updated to use this.