manifest: rename TP roleBinding to cluster-monitoring-operator-alert-customization
Problem: When the alertoverrides feature was promoted to v1, I failed to
rename a RoleBinding accordingly. This caused CVO to fail to progress
through an update since it had to modify an existing RoleBinding (cannot
change RoleRef).

Solution: Rename the RoleBinding object in manifests/. We don't need to
mark the old object for deletion, since it was behind a tech-preview
flag, and tech-preview cluster are not supported to upgrade. So from the
upgrade process this is simply a new resource.

Links: https://issues.redhat.com/browse/OCPBUGS-15210

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>