OCPBUGS-14922: skip console-plugin installation if console capability is disabled
This PR fixes the failure to upgrade from OpenShift version 4.13 to
4.14 in environments (such as telco) where `console` capability is
disabled. The upgrade fails  when CMO executes the console-plugin task
which fails due to missing ConsolePlugin CRD.

The fix checks if `console` capability is enabled in ClusterVersion -
version and bails out (with a log message) if it is disabled.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>