OCPBUGS-15440: fix CMO to apply console-plugin pod.spec config
The console-plugin's deployment configs that need to applied to the
deployment.spec.template.spec weren't getting applied because the
changes were made to the copy of the object instead of the original
object itself. The tests didn't catch the issue because the pod assertions
didn't validate if any pods were found for the combination of namespace
and label-selector

This commit fixes both by refering (pointer) to the original spec
instead of the copy and by fixing the assertions fail if it fails to
find any pods.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>