pkg/client: Only update Prometheus if we have changes to push
Following the pattern we've used in CreateOrUpdateDeployment since
675a7ed490 (pkg/*: add k8s functions for Telemeter client, 2018-09-21, #103).
This saves some network traffic and Kubernetes API service load.
There's a bit of dancing as I copy Status (which is irrelevant for
this use-case) and ObjectMeta (except for merged annotations and
labels) over from 'existing' to 'required', but that sets up a
convenient DeepEqual for "did anything we care about change?".

It also makes it easier to see from the Kube API server logs
when the Prometheus resources are actually being updates, while before
this commit:

  $ curl -s https://gcsweb-ci.apps.ci.l2s4.p1.openshiftapps.com/gcs/origin-ci-test/logs/periodic-ci-openshift-release-master-nightly-4.14-e2e-aws-ovn-serial/1685027676433158144/artifacts/e2e-aws-ovn-serial/gather-audit-logs/artifacts/audit-logs.tar | tar -xz --strip-components=2
  $ zgrep -h '"resource":"prometheuses"' kube-apiserver/*.log.gz | jq -r 'select(.verb == "update" and .objectRef.subresource != "status") | .stageTimestamp + " " + (.responseStatus.code | tostring) + " " + .user.username' | sort

would find traffic like:

  2023-07-28T21:10:30.455712Z 200 system:serviceaccount:openshift-monitoring:cluster-monitoring-operator
  2023-07-28T21:11:39.629004Z 200 system:serviceaccount:openshift-monitoring:cluster-monitoring-operator
  2023-07-28T21:11:58.727870Z 200 system:serviceaccount:openshift-monitoring:cluster-monitoring-operator
  2023-07-28T21:13:24.616877Z 200 system:serviceaccount:openshift-monitoring:cluster-monitoring-operator
  2023-07-28T21:13:43.859596Z 200 system:serviceaccount:openshift-monitoring:cluster-monitoring-operator
  2023-07-28T21:14:51.770214Z 200 system:serviceaccount:openshift-monitoring:cluster-monitoring-operator
  2023-07-28T21:15:10.524179Z 200 system:serviceaccount:openshift-monitoring:cluster-monitoring-operator
  ...

I've also added diff logging, using an already-vendored library, to
make it easier to understand why the operator feels the need to update
the resource.

The go.mod update was generated with:

  $ go mod tidy

using:

  $ go version
  go version go1.19.5 linux/amd64

now that we're directly using the already-vendored package.