jsonnet/rules: exclude -1 from etcd objects count
The apiserver_storage_objects metric can take -1 values when there are
communication issues between the kube-apiserver and etcd. In such case,
the kube-apiserver can't get objects count from etcd and as such exposes
a metric value denoting the issue. However, this is polluting the data
exported to telemetry so they should be skipped. Doing that will not
cause the lose of the etcd issue information since we have alerts that
would trigger in such scenarios.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>