MON-3479: update Prometheus operator assets to v0.69.1
This commit also pins the prometheus-operator jsonnet to the current
stable release branch of the operator. The reason is that pinning to the
main branch has the drawback of pulling CRD changes which are not yet
supported by the downstream operator.

The commit also updates the different jsonnet dependencies to their
latest versions and makes adjustments to the CMO jsonnet to be
compatible with OCP requirements (in particular to the security contexts
of node_exporter and thanos-querier).

Signed-off-by: Simon Pasquier <spasquie@redhat.com>