OCPBUGS-25378: drop InfoInhibitor from default Alertmanager config
OCP doesn't include the InfoInhibitor alerting rule out of the box. To
avoid confusion, it is desired to remove the alert from the default
Alertmanager configuration.

Note that the change will only apply to newly created clusters because
the Cluster Monitoring Operator only handles the creation of the
Alertmanager configuration secret and it never updates when the secret
already exists.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>