MON-3701: clean-up trusted CA bundle for Alertmanager
CMO needed to inject the hashed version of the trusted CA bundle
ConfigMap into the Alertmanager resource when OAuth proxy was used
because it couldn't detect updates to the bundle and reload it.

The trusted CA bundle is still required to be mounted into the pod for
the Alertmanager container but it is declared directly in the static
manifest instead of being injected at runtime by CMO since Alertmanager
reloads the CA whenever it changes.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>