OCPBUGS-32510: change metrics-server probes for SNO
This change switches the metrics-server's readiness probe to use the
`/livez` endpoint instead of `/readyz` for single-node deployments.

It also adds a startup probe using the same characteristics as the
default readiness probe to ensure that the pod reports ready only when
it has gathered enough samples from kubelet.

By default, the `/readyz` endpoint is used to assert the component
readiness. This endpoint returns success when the metrics-server has
metric samples over 2 intervals (e.g. it has scraped at least one
kubelet twice).

In single-node deployments, it happens sometimes (especially in
end-to-end tests) that the kubelet fails to respond in a timely fashion
due to contention in cAdvisor, leading to a delayed readiness (and test
failures). To workaround the issue, we use the `/livez` endpoint in this
mode.

The long-term plan is to switch resource metrics from cAdvisor to the
CRI stats API (currently an alpha feature). Once it happens, we can
remove this change.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>