Allow scrapping of OpenStack load-balancer-id
In order to enable correlated metrics between OpenShift and the
underlying platform, we need a way to map resources. We already have it
for instances and volumes via the `kube_node_info` and
`kube_persistentvolume_info` respectively, however we don't have the
load-balancer ID in `kube_service_info`.

For OpenStack, this mapping lives in the
`loadbalancer.openstack.org/load-balancer-id` annotation in the service
object.

This patches enables scrapping OpenStack load-balancer-id, which can be
found in the metric `kube_service_annotations`, label
`annotation_loadbalancer_openstack_org_load_balancer_id`.