Always generate CRD metrics
When the VPA metrics were (re-)added, code was added to not try to
watch VPAs unless the CRDs had been created, to avoid spurious errors.
But it appears that ksm has since been fixed to not need this; it
watches for CRD creation/deletion itself, and ignores the VPA metric
definitions if the VPA CRDs don't exist.

Signed-off-by: Dan Winship <danwinship@redhat.com>