chore(jsonnet): use prometheus_remote_storage_queue_highest_timestamp_in_seconds in PrometheusRemoteWriteBehind
This metric was introduced in https://github.com/openshift/prometheus/pull/262 and related PRs.

Dashboard expressions are not changed, since updating them may be more complex. Fixing the alert is more important and we can always revisit that if it causes confusion.

On main, dashboards will be adjusted later once the jsonnet dependencies are updated.