MON-4420: Enable Gateway API collection by telemeter
This change enables the collection of Gateway API count metric by telemeter,
providing insights of the usage of Gateway API on Openshift.

The generated metric contains only a count of the amount of created Gateway resources,
aggregated by the Gateway Class. In case the Gateway Class is the one managed by
Openshift, the metric will have a label 'gateway_class_type' with value 'openshift'.

In case the Gateway uses a Gateway Class that is not managed by Openshift Gateway API
offer, the label 'gateway_class_type' will contain the value 'not-openshift'.